setonix singularity module : other fix

+++ setonix/setup_scripts/setup_singularity_pawsey_modules.sh
@@ -20,18 +20,22 @@ dst_dir="${root_dir}/${singularity_symlink_module_dir}"
 
 # ensure destination directory exists
 mkdir -p ${dst_dir}
+# remove old pawsey singularity modules
+rm -f ${dst_dir}/*.lua
 
-for version in $( ls ${src_dir} ) ; do
+for version in $( ls ${src_dir}/*.lua ${src_dir}/.*.lua 2>/dev/null ) ; do
+  #2. hide Spack module
+  if [ -e ${src_dir}/${version} ] ; then
+    mv ${src_dir}/${version} ${src_dir}/.${version}
+  fi
   # 1.B singularity-astro is just the original module
-  cp -p ${src_dir}/${version} ${dst_dir}/${version/\.lua/-astro.lua}
+  cp -p ${src_dir}/.${version} ${dst_dir}/${version/\.lua/-astro.lua}
   # 1.A singularity does not bind mount /askapbuffer and /astro
   sed \
     -e '/SINGULARITY_BINDPATH/ s;/askapbuffer;;g' \
     -e '/SINGULARITY_BINDPATH/ s;/astro;;g' \
     ${dst_dir}/${version/\.lua/-astro.lua} \
     >${dst_dir}/${version}
-  #2. hide Spack module
-  mv ${src_dir}/${version} ${src_dir}/.${version}
 done
